test(mcp): rework compression tests for zarrista backend
Rename test_tensorstore_compression.py to test_zarrista_compression.py:
the blosc, gzip, and zarr v2/v3 coverage now runs through the default
zarrista write path (no use_tensorstore flag, no tensorstore
importorskips) and each test reads the first scale back with
zarr-python to verify other zarr implementations can decode the
compressed chunks zarrista wrote.

Add test_use_tensorstore_alias_contract: a conversion with
use_tensorstore=True must still succeed, produce a store that passes
validate_ome_zarr, and not leak the DeprecationWarning out of the MCP
tool (enforced via a filterwarnings=error marker).

Audit of test_convert_to_ome_zarr.py and test_basic.py found no
store-object inputs or zarr-python-specific attrs assumptions; the
zarr.open verification read is kept as the compatibility check of
zarrista-written output, now with a comment noting zarr-python is a
test-only dependency.
